Complex numbers almost work
This commit works in the case of the cameraman ptychography
dataset that requires position updates, if you use delta_beta
and Adam Optimizer.

The real_imag result needs to be debugged, and the other optimizers
need to be evaluated - they previously relied on the derivatives
being real valued and that assumption is now incorrect. So
optimizers.py needs to be fixed.